+++ selene-lib/src/rules/parenthese_conditions.rs
@@ -54,9 +54,9 @@ impl ParentheseConditionsVisitor {
     fn lint_condition(&mut self, condition: &ast::Expression) {
         let is_parentheses = match condition {
             ast::Expression::Parentheses { .. } => true,
-            ast::Expression::Value { value, .. } => {
+            ast::Expression::Value { value, binop } => {
                 if let ast::Value::ParseExpression(_) = &**value {
-                    true
+                    binop.is_none()
                 } else {
                     false
                 }