+++ selene-lib/tests/lints/parenthese_conditions/parenthese_conditions.lua
@@ -11,3 +11,5 @@ while (true) do end
 
 repeat until false
 repeat until (false)
+
+while (true) and false do end